Fixed bugs and added selfmade test feature

+++ src/main/java/utils/GCD.java
@@ -5,32 +5,33 @@ import org.slf4j.LoggerFactory;
 
 public class GCD {
     private static Logger logger = LoggerFactory.getLogger(GCD.class);
-    private Integer counter;
-    private Integer divisor;
-    private Integer result;
-    private Integer[] x;
-    private Integer[] y;
-    private Integer iteration;
+    private long counter;
+    private long divisor;
+    private long result;
+    private long[] x;
+    private long[] y;
+    private long iteration;
 
-    public GCD(Integer counter, Integer divisor) {
+    public GCD(long counter, long divisor) {
         this.counter = counter;
         this.divisor = divisor;
         result = 0;
         iteration = 1;
-        x = new Integer[]{1, 0, 0};
-        y = new Integer[]{0, 1, 0};
+        x = new long[]{1, 0, 0};
+        y = new long[]{0, 1, 0};
     }
 
-    public Integer calculateGCD() {
+    public long calculateGCD() {
         return calculate();
     }
 
-    private Integer calculate() {
+    private long calculate() {
         logger.info("");
+
         while (true) {
             result = counter / divisor;
-            Integer newCounter = divisor;
-            Integer previousDivisor = divisor;
+            long newCounter = divisor;
+            long previousDivisor = divisor;
             divisor = counter % divisor;
 
             if (divisor == 0) {
@@ -51,12 +52,12 @@ public class GCD {
         }
     }
 
-    private void countNewXY(Integer[] x, Integer[] y) {
+    private void countNewXY(long[] x, long[] y) {
         x[2] = x[1] * result + x[0];
         y[2] = y[1] * result + y[0];
     }
 
-    private void shiftXY(Integer[] x, Integer[] y) {
+    private void shiftXY(long[] x, long[] y) {
         x[0] = x[1];
         x[1] = x[2];
         x[2] = 0;
@@ -66,21 +67,21 @@ public class GCD {
         y[2] = 0;
     }
 
-    public Integer getGCD() {
+    public long getGCD() {
         return divisor;
     }
 
-    public Integer getX() {
+    public long getX() {
         x[2] = (int) (x[2] * Math.pow(-1, iteration));
         return x[2];
     }
 
-    public Integer getY() {
+    public long getY() {
         y[2] = (int) (y[2] * Math.pow(-1, iteration + 1));
         return y[2];
     }
 
-    public Integer getIteration() {
+    public long getIteration() {
         return iteration;
     }
 }
